requirements.txt
absl-py==0.9.0
alabaster==0.7.12
anaconda-client==1.7.2
anaconda-navigator==1.9.7
anaconda-project==0.8.3
appdirs==1.4.3
asn1crypto==1.0.1
astor==0.8.1
astroid==2.3.1
astropy==3.2.2
atomicwrites==1.3.0
attrs==19.2.0
Babel==2.7.0
backcall==0.1.0
backports.functools-lru-cache==1.6.1
backports.os==0.1.1
backports.shutil-get-terminal-size==1.0.0
backports.tempfile==1.0
backports.weakref==1.0.post1
beautifulsoup4==4.8.0
bitarray==1.0.1
bkcharts==0.2
bleach==3.1.0
blosc==1.9.1
bokeh==1.3.4
boto==2.49.0
Bottleneck==1.2.1
Brian2==2.3.0.2
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
Click==7.0
cloudpickle==1.2.2
clyent==1.2.2
colorama==0.4.1
ConceptNet==5.7.0
conda==4.8.4
conda-build==3.18.9
conda-package-handling==1.7.0
conda-verify==3.4.2
ConfigArgParse==1.0
contextlib2==0.6.0
cryptography==2.7
cycler==0.10.0
Cython==0.29.13
cytoolz==0.10.0
dask==2.5.2
datrie==0.8
decorator==4.4.0
defusedxml==0.6.0
distlib==0.3.0
distributed==2.5.2
docutils==0.15.2
easydict==1.9
elephant==0.8.0
entrypoints==0.3
et-xmlfile==1.0.1
exdir==0.4.1
fastcache==1.1.0
filelock==3.0.12
Flask==1.1.1
fsspec==0.5.2
ftfy==5.7
future==0.18.2
gast==0.3.2
gevent==1.4.0
gitdb==4.0.2
GitPython==3.1.0
glob2==0.7
gmpy2==2.0.8
google-pasta==0.1.8
greenlet==0.4.15
grpcio==1.26.0
h5py==2.10.0
hdbscan==0.8.26
HeapDict==1.0.1
herdingspikes==0.3.7
html5lib==1.0.1
idna==2.8
imageio==2.6.0
imagesize==1.1.0
importlib-metadata==0.23
ipykernel==5.1.2
ipython==7.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
isosplit5==0.1.3
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.15.1
jeepney==0.4.1
Jinja2==2.10.3
joblib==0.13.2
jp-proxy-widget==1.0.7
json5==0.8.5
jsonschema==3.0.2
jupyter==1.0.0
jupyter-client==5.3.3
jupyter-console==6.0.0
jupyter-core==4.5.0
jupyter-ui-poll==0.1.2
jupyterlab==1.1.4
jupyterlab-server==1.0.6
kachery==0.6.5
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==18.0.0
kiwisolver==1.1.0
klusta==3.0.16
klustakwik2==0.2.7
langcodes==1.4.1
lazy-object-proxy==1.4.2
lazyarray==0.3.3
LFPy==2.1
libarchive-c==2.8
lief==0.9.0
llvmlite==0.29.0
locket==0.2.0
lxml==4.4.1
marisa-trie==0.7.5
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mccabe==0.6.1
# Editable install with no version control (MEArec==1.4.3)
-e /home/alex/anaconda3/lib/python3.7/site-packages
MEAutility==1.4.7
mistune==0.8.4
mkl-fft==1.0.14
mkl-random==1.1.0
mkl-service==2.3.0
ml-ms4alg==0.3.5
mock==3.0.5
more-itertools==7.2.0
mountainlab-pytools==0.7.5
mpi4py==3.0.3
mpmath==1.1.0
msgpack==0.6.1
msgpack-python==0.5.6
multipledispatch==0.6.0
natsort==7.0.1
navigator-updater==0.2.1
nbconvert==5.6.0
nbformat==4.4.0
neo==0.8.0
networkx==2.3
NEURON==7.7
nltk==3.4.5
nose==1.3.7
notebook==6.0.1
numba==0.45.1
numexpr==2.7.0
numpy==1.17.2
numpydoc==0.9.1
olefile==0.46
openpyxl==3.0.0
ordered-set==3.1.1
packaging==19.2
pandas==0.25.1
pandocfilters==1.4.2
parso==0.5.1
partd==1.0.0
path.py==12.0.1
pathlib2==2.3.5
patsy==0.5.1
pep8==1.7.1
pexpect==4.7.0
pickle-mixin==1.0.2
pickleshare==0.7.5
Pillow==6.2.0
pkginfo==1.5.0.1
pluggy==0.13.0
ply==3.11
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.11.2
psutil==5.6.3
psycopg2-binary==2.8.4
ptyprocess==0.6.0
py==1.8.0
pybind11==2.5.0
pycodestyle==2.5.0
pycosat==0.6.3
pycparser==2.19
pycrypto==2.6.1
pycurl==7.43.0.3
pyflakes==2.1.1
Pygments==2.4.2
pyintan==0.1.6
pylint==2.4.2
pymongo==3.11.0
PyNN==0.9.5
pyodbc==4.0.27
pyopenephys==1.0.7
pyOpenSSL==19.0.0
pyparsing==2.4.2
pyqtgraph==0.10.0
pyrsistent==0.15.4
PySocks==1.7.1
pytest==5.2.1
pytest-arraydiff==0.3
pytest-astropy==0.5.0
pytest-doctestplus==0.4.0
pytest-openfiles==0.4.0
pytest-remotedata==0.3.2
python-dateutil==2.8.0
python-frontmatter==0.5.0
pytz==2019.3
PyWavelets==1.0.3
PyYAML==5.1.2
pyzmq==18.1.0
QtAwesome==0.6.0
qtconsole==4.5.5
QtPy==1.9.0
quantities==0.12.4
ratelimiter==1.2.0.post0
regex==2018.2.21
requests==2.22.0
rope==0.14.0
ruamel-yaml==0.15.46
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
sagemath==1.3.0
scanimage-tiff-reader==1.4.1
scikit-image==0.15.0
scikit-learn==0.23.2
scipy==1.3.1
seaborn==0.9.0
SecretStorage==3.1.1
Send2Trash==1.5.0
simplegeneric==0.8.1
simplejson==3.17.2
singledispatch==3.4.0.3
six==1.12.0
sklearn==0.0
smmap==3.0.1
snakemake==5.10.0
snowballstemmer==2.0.0
sortedcollections==1.1.2
sortedcontainers==2.1.0
soupsieve==1.9.3
Sphinx==2.2.0
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sphinxcontrib-websupport==1.1.2
spikecomparison==0.3.0
spikeextractors==0.8.4
spikefeatures==0.1.1
-e git+https://github.com/flatironinstitute/spikeforest2.git@b18e6407e5e2b8582d1901a5e884bd6514646a1e#egg=spikeforest2
spikeinterface==0.9.9
spikemetrics==0.2.2
spikesorters==0.4.1
spiketoolkit==0.7.0
spikewidgets==0.5.0
spyder==3.3.6
spyder-kernels==0.5.2
spyking-circus==0.9.9
SQLAlchemy==1.3.9
statsmodels==0.10.1
style==1.1.0
suite2p==0.7.1
sympy==1.4
tables==3.5.2
tblib==1.4.0
tensorboard==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
threadpoolctl==2.1.0
toolz==0.10.0
toposort==1.5
torch==1.4.0
torchvision==0.5.0
tornado==6.0.3
tqdm==4.36.1
traitlets==4.3.3
tridesclous==1.6.0
unicodecsv==0.14.1
update==0.0.1
urllib3==1.24.2
vdom==0.6
virtualenv==20.0.25
ViSAPy==1.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
wordfreq==2.2.1
wrapt==1.11.2
wurlitzer==1.0.3
xlrd==1.2.0
XlsxWriter==1.2.1
xlwt==1.3.0
xmltodict==0.11.0
zict==1.0.0
zipp==0.6.0
